Add a variation of the "sonic weapon is not in your hands" message.
Two versions:

(1)

You put a sonic lance in your backpack.
...
Your sonic weapon is not in your hands.

(2)

The Vvrael warlock stows a midnight black spiked whip to pick up a sonic
lance, but the lance disappears!  With a scowl, the warlock draws forth
his whip.
...
Your sonic weapon is not in your hands.  You are unable to renew the
song.