Fix muckled? function.
The muckled? function uses 'or' instead of '||'. This causes it, in effect, to only ever check for being webbed as the 'or' operator has lower precedence than the assignment operator. e.g. muckled = checkwebbed or checkdead... is reads as (muckled = checkwebbed) or checkdead... The rest isn't processed since muckled = checkwebbed succeeds.